added tests for fetch add ons and fetch user backings query.
